perf: reuse OpenAI client and add undici keep-alive Agent with connection warmup
Extract OpenAI client creation logic into src/common/openai-client.ts:
- Custom undici Agent with 60s keepAlive timeout (default is 4s)
- Module-level client instance cache (reuse across calls)
- Fire-and-forget connection warmup on first creation (3s timeout)
- getMachineId() helper

The App.tsx now simply imports and re-exports createOpenAIClient from
the new common module, keeping UI concerns separate from HTTP/client
lifecycle management.
